Q{5}
Now create a more sophisticated Rectangle class than the one you created in problem 4. This
class stores only the Cartesian coordinates of the four corners of the rectangle. The constructor
calls a set function that accepts four sets of coordinates and verifies that each of these is in the
first quadrant with no single x- or y-coordinate larger than 20.0. The set function also verifies
that the supplied coordinates do, in fact, specify a rectangle. Provide member functions that
calculate the length, width, perimeter and area. The length is the larger of the two dimensions.
Include a predicate function square that determines whether the rectangle is a square.